# crates/paddock-runner/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paddock Runner API",
    "version": "__VERSION__",
    "description": "The complete HTTP surface of one Paddock runner (one served model, one port).\n\nThe surface has three families:\n\n- **OpenAI-compatible** - `/v1/chat/completions`, `/v1/completions`, `/v1/responses`, `/v1/embeddings`, `/v1/audio/transcriptions`, `/v1/models`. These conform to OpenAI's API and are documented by OpenAI's own reference; this document lists them with Paddock's deviations and extensions only, and deliberately does not re-specify their schemas. Point an official OpenAI SDK at the runner's base URL and it works.\n- **Anthropic-compatible** - `/v1/messages`, `/v1/messages/count_tokens`. Same rule: Anthropic's reference is the specification.\n- **Paddock-native** - forced alignment, rerank, the extraction preview, the model card, stats, and the realtime socket. These exist nowhere else, so they are fully specified here.\n\nAuthentication: a runner started without an API key accepts unauthenticated requests. With a key configured, `/v1/*` and `/api/*` require `Authorization: Bearer <key>`; `/healthz` and this document stay open, and `/metrics` is open on loopback but keyed over the network.\n\nOne runner serves one model, so `model` request fields are accepted and ignored everywhere; `GET /v1/models` lists exactly one entry with Paddock capability extensions.",
    "license": { "name": "MIT OR Apache-2.0", "identifier": "MIT OR Apache-2.0" }
  },
  "servers": [
...
  "security": [{}, { "bearerAuth": [] }],
...
  "paths": {
...
    "/api/extract": {
      "post": {
        "tags": ["paddock"],
        "summary": "Extraction preview: what the model would read from an attachment",
        "description": "Runs the same extraction lanes a chat send runs (PDF text, docx, spreadsheets, text-native decode, photo metadata) and returns the injection text verbatim - metadata block included - so a client can show exactly what a prompt will carry. Refusals (encrypted PDF, undecodable bytes) are the same honest 400s a send would produce.",
        "operationId": "extractPreview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "data": { "type": "string", "description": "The file as a data URI or bare base64" },
                  "filename": { "type": "string", "description": "Original filename, used for type detection" },
                  "file_metadata": { "type": "string", "enum": ["on", "off"], "description": "Whether the metadata block rides along (default on) - same contract as the inference surfaces" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The extraction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["text", "kind"],
                  "properties": {
                    "text": { "type": "string", "description": "The injection text, verbatim" },
                    "kind": { "type": "string", "description": "Which extraction lane ran" },
                    "pages": { "type": ["integer", "null"], "description": "Page count where the lane has one" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/metadata": {
      "post": {
        "tags": ["paddock"],
        "summary": "Embedded file metadata: everything the file says about itself",
        "description": "Every metadata field a file carries, grouped by source - EXIF, MakerNotes, XMP, IPTC, ICC, PDF Info, QuickTime and derived Composite tags for photos and documents; core.xml, app.xml and custom.xml properties for Office packages. Distinct from /api/extract, which answers what the MODEL would read and is shaped by this server's context window and vision capability: this one depends on nothing but the bytes. A file that carries no metadata is an ordinary empty answer, not an error.",
        "operationId": "fileMetadata",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "data": { "type": "string", "description": "The file as a data URI or bare base64" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["format", "reader", "groups"],
                  "properties": {
                    "format": { "type": ["string", "null"], "description": "What the BYTES are (\"JPEG\", \"PDF\", \"Office Open XML\" ...), never the declared mime; null when the magic bytes match nothing we read" },
                    "reader": { "type": "string", "enum": ["sift", "scriptor", "none"], "description": "Which reader answered" },
                    "groups": {
                      "type": "array",
                      "description": "Groups in reader order; empty when the file carries no metadata",
                      "items": {
                        "type": "object",
                        "required": ["name", "tags"],
                        "properties": {
                          "name": { "type": "string", "description": "An open set, named by the reader and the file: EXIF, MakerNotes, XMP, IPTC, ICC, PDF, QuickTime, HEIF, Composite, File - or Document, Application, Custom for Office packages. A group is a namespace: the same name in two groups is two sources stating it, not a duplicate." },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": ["name", "value"],
                              "properties": {
                                "name": { "type": "string" },
                                "value": { "type": "string", "description": "Display-ready; control characters collapse to spaces" },
                                "truncated": { "type": "boolean", "description": "Present and true when the value hit the 4096-character ceiling" }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required on /v1/* and /api/* only when the runner was configured with an API key; a keyless local runner accepts unauthenticated requests."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request - OpenAI-shaped error body with the refusal spelled out",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The OpenAI-shaped error envelope every surface uses, 404s included",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "type"],
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string", "description": "e.g. invalid_request_error, not_found_error, internal_error" },
              "param": { "type": ["string", "null"] },
              "code": { "type": ["string", "null"] }
            }
          }
        }
      },
...
    }
  }
}